requirements.txt
digi-xbee==1.4.0